// test/src/pages/sandbox/users/ichibasan/sandbox-builder/views/AddContentView/types.ts
export type ComponentKey =
  | "Accordion"
  | "ActionList"
  | "Avatar"
  | "AvatarGroup"
  | "Badge"
  | "Banner"
  | "Blockquote"
  | "Breadcrumb"
  | "Button"
  | "ButtonGroup"
  | "Calendar"
  | "Card"
  | "Checkbox"
  | "CheckboxCard"
  | "CheckboxGroup"
  | "Code"
  | "CodeBlock"
  | "Combobox"
  | "ContentHeader"
  | "DataTable"
  | "DateField"
  | "DatePicker"
  | "DescriptionList"
  | "Divider"
  | "DividerVertical"
  | "EmptyState"
  | "FileDrop"
  | "Form"
  | "FormControl"
  | "IconButton"
  | "InformationCard"
  | "InformationCardGroup"
  | "Link"
  | "Mark"
  | "NavList"
  | "OrderedList"
  | "Pagination"
  | "Radio"
  | "RadioCard"
  | "RadioGroup"
  | "RangeCalendar"
  | "RangeDateField"
  | "RangeDatePicker"
  | "RangeTimeField"
  | "RangeTimePicker"
  | "Search"
  | "SegmentedControl"
  | "Select"
  | "Skeleton"
  | "StatusLabel"
  | "Stepper"
  | "Switch"
  | "Table"
  | "Tabs"
  | "Tag"
  | "TagGroup"
  | "TagInput"
  | "TagPicker"
  | "Text"
  | "TextField"
  | "Textarea"
  | "TimeField"
  | "TimePicker"
  | "Timeline"
  | "Toolbar"
  | "Tree"
  | "UnorderedList"
  | "SideNavigation";

export type ContentItem = {
  id: string;
  component: ComponentKey;
  props?: Record<string, string>;
  slot?: "start" | "end";
};